clipping-logic/requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
cssselect==1.2.0
cssselect2==0.7.0
cssutils==2.9.0
csutils==0.1.9
docx==0.2.4
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.0
markdown-it-py==3.0.0
markdown2==2.4.13
markdownify==0.12.1
MarkupSafe==2.1.5
mdpdf==0.0.18
mdurl==0.1.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pandoc==2.3
pdfkit==1.0.0
pdflatex==0.1.3
pillow==10.2.0
plumbum==1.8.2
ply==3.11
premailer==3.10.0
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.22
pydeck==0.8.1b0
pydyf==0.9.0
Pygments==2.17.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypandoc==1.13
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
referencing==0.34.0
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
selenium==4.19.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.33.0
tenacity==8.2.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
weasyprint==61.2
webencodings==0.5.1
wkhtmltopdf==0.2
wsproto==1.2.0
zopfli==0.2.3
